+++ Clicked/Locales/enUS.lua
@@ -309,3 +309,5 @@ L["Automatically import from macros"] = true
 L["Are you sure you want to delete these %d bindings and/or groups?"] = true
 L["copy"] = true
 L["Action bar page"] = true
+L["Enter the bonus bar page number."] = true
+L["Enter the action bar page number."] = true